BACKWARD COMPATABILTY LOST: separate initaliser file for the reduced gravity functionality
